Speed improvement by explicit floats in projection

+++ MATLAB/Source/ray_interpolated_projection_parallel.cu
@@ -134,9 +134,9 @@ __global__ void kernelPixelDetector_parallel( Geometry geo,
     S.z=(source.z+pixelU*deltaU.z+pixelV*deltaV.z);
     
     // Length is the ray length in normalized space
-    double length=sqrt((S.x-P.x)*(S.x-P.x)+(S.y-P.y)*(S.y-P.y)+(S.z-P.z)*(S.z-P.z));
+    double length=sqrtf((S.x-P.x)*(S.x-P.x)+(S.y-P.y)*(S.y-P.y)+(S.z-P.z)*(S.z-P.z));
     //now legth is an integer of Nsamples that are required on this line
-    length=ceil(length/geo.accuracy);//Divide the directional vector by an integer
+    length=ceilf(length/geo.accuracy);//Divide the directional vector by an integer
     vectX=(P.x -S.x)/(length);
     vectY=(P.y -S.y)/(length);
     vectZ=(P.z -S.z)/(length);
@@ -150,10 +150,10 @@ __global__ void kernelPixelDetector_parallel( Geometry geo,
     
     // limit the amount of mem access after the cube, but before the detector.
     if ((2*DSO/geo.dVoxelX+maxdist)/geo.accuracy  <   length)
-        length=ceil((2*DSO/geo.dVoxelX+maxdist)/geo.accuracy);  
+        length=ceilf((2*DSO/geo.dVoxelX+maxdist)/geo.accuracy);  
     //Length is not actually a length, but the amount of memreads with given accuracy ("samples per voxel")
     
-    for (i=floor(maxdist/geo.accuracy); i<=length; i=i+1){
+    for (i=floorf(maxdist/geo.accuracy); i<=length; i=i+1){
         tx=vectX*i+S.x;
         ty=vectY*i+S.y;
         tz=vectZ*i+S.z;
@@ -161,8 +161,9 @@ __global__ void kernelPixelDetector_parallel( Geometry geo,
         sum += tex3D(tex, tx+0.5, ty+0.5, tz+0.5); // this line is 94% of time.
         
     }
-    float deltalength=sqrt((vectX*geo.dVoxelX)*(vectX*geo.dVoxelX)+
-            (vectY*geo.dVoxelY)*(vectY*geo.dVoxelY)+(vectZ*geo.dVoxelZ)*(vectZ*geo.dVoxelZ) );
+    float deltalength=sqrtf((vectX*geo.dVoxelX)*(vectX*geo.dVoxelX)+
+                            (vectY*geo.dVoxelY)*(vectY*geo.dVoxelY)+
+                            (vectZ*geo.dVoxelZ)*(vectZ*geo.dVoxelZ) );
     detector[idx]=sum*deltalength;
 }
